Get things working again.
Add prettier and eslint and clean the code up with modern JavaScript.

Add an .editorconfig file.

Update dependencies.

Also add a development server that serves a local html file for testing.

Now it is expected that the tikz code be wrapped in
`\begin{tikzpicture}` and `\end{tikzpicture}` and the
`data-tikz-options` option removed.  This means that other environments
can also be used instead of a `tikzpicture` environment.