Change to using the opentype cmex fonts.
The glyph for the angbracketleftbig character in the truetype cmex fonts
is a soft hyphen which is a character that browsers will not show. As
such the character will not be seen if used.  For example, if
`\big\langle` is used in TeX. Note that `\left\langle` could result in
this character as well.

To complicate matters the otf fonts have invalid out of range glyphs in
them.  So those glyphs need to be removed.  The opentype fonts are
used converted into woff2 fonts since the library used does not support
writing opentype fonts. Also, some browsers may not support opentype
fonts.